refactor: Move annotations to data.annotation
