+++ src/winter/data/annotation/verb/POST.java
@@ -1,4 +1,4 @@
-package winter.annotation.classlevel;
+package winter.data.annotation.verb;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -6,6 +6,6 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.ElementType;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface Controller {
+@Target(ElementType.METHOD)
+public @interface POST {
 }
